+++ examples/nextjs/js/sample-app/src/app/layout.js
@@ -17,7 +17,7 @@ export default function RootLayout({ children }) {
           {`
             (function() {
               "use strict";
-              window.RudderSnippetVersion = "3.0.9";
+              window.RudderSnippetVersion = "3.0.10";
               var identifier = "rudderanalytics";
               if (!window[identifier]) {
                 window[identifier] = [];