+++ packages/analytics-js-common/src/types/DataResidency.ts
@@ -1,9 +1,0 @@
-export type RegionDetails = {
-  url: string;
-  default: boolean;
-};
-
-/**
- * Represents residency server input the options
- */
-export type ResidencyServerRegion = 'US' | 'EU';